% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.ResourceClient}
\alias{as.data.frame.ResourceClient}
\title{Coerce a ResourceClient object to a data.frame}
\usage{
\method{as.data.frame}{ResourceClient}(x, ...)
}
\arguments{
\item{x}{a ResourceClient object}

\item{...}{additional parameters, that may be used (or ignored) by the resource client.}
}
\value{
a data.frame (or a tibble)
}
\description{
Attempt to coerce a resource object to a data.frame: find a ResourceResolver and get the
ResourceClient that will connect to the described dataset and make a data.frame of it.
}
