% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.R
\name{as.resource.object}
\alias{as.resource.object}
\title{Coerce resource client to the internal data object}
\usage{
as.resource.object(x)
}
\arguments{
\item{x}{The ResourceClient object to coerce to a data.frame}
}
\description{
Coerce a ResourceClient object to internal data object: depending on the implementation of the ResourceClient,
it can be a data connection object (like a DBI connection to a SQL database),
or the actual data structure (when a resource is a R object extracted from a R data file for instance).
}
