% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.R
\name{as.resource.tbl}
\alias{as.resource.tbl}
\title{Coerce resource client to a tbl}
\usage{
as.resource.tbl(x, ...)
}
\arguments{
\item{x}{The ResourceClient object to coerce to a data.frame}

\item{...}{Additional parameters, that may be used (or ignored) by the resource client.}
}
\value{
a dplyr's tbl
}
\description{
Coerce a ResourceClient object to a dplyr's tbl.
}
