% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandResourceClient.R
\docType{class}
\name{CommandResourceClient}
\alias{CommandResourceClient}
\title{Command resource client}
\format{A R6 object of class CommandResourceClient}
\usage{
CommandResourceClient
}
\description{
Base class for resource clients issuing commands and get a result with the status of the execution,
the output and the error messages.
}
\keyword{datasets}
