% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIResourceConnector.R
\docType{class}
\name{DBIResourceConnector}
\alias{DBIResourceConnector}
\title{DBI resource connector}
\format{A R6 object of class DBIResourceConnector}
\usage{
DBIResourceConnector
}
\description{
Makes a DBI connection from a resource description, used in SQLResourceClient that is based on DBI.
}
\section{Methods}{


\code{$new()} Create new DBIResourceConnector instance.
\code{$isFor(resource)} Get a logical that indicates that the DBI connector is applicable to the provided resource object.
\code{$createDBIConnection(resource, ...)} Get the DBI connection object described by the provided resource.
\code{$closeDBIConnection(conn)} Release the DBI connection when done.
}

\keyword{datasets}
