% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DremioResourceConnector.R
\docType{class}
\name{DremioResourceConnector}
\alias{DremioResourceConnector}
\title{Dremio DBI resource connector}
\format{
A R6 object of class DremioResourceConnector
}
\description{
Dremio DBI resource connector

Dremio DBI resource connector
}
\details{
Makes a ODBC DBI connection to a Dremio server from a resource description.
}
\section{Super classes}{
\code{\link[resourcer:DBIResourceConnector]{resourcer::DBIResourceConnector}} -> \code{\link[resourcer:ODBCResourceConnector]{resourcer::ODBCResourceConnector}} -> \code{DremioResourceConnector}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DremioResourceConnector$new()}}
\item \href{#method-isFor}{\code{DremioResourceConnector$isFor()}}
\item \href{#method-getConnectionString}{\code{DremioResourceConnector$getConnectionString()}}
\item \href{#method-clone}{\code{DremioResourceConnector$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="resourcer" data-topic="DBIResourceConnector" data-id="closeDBIConnection">}\href{../../resourcer/html/DBIResourceConnector.html#method-closeDBIConnection}{\code{resourcer::DBIResourceConnector$closeDBIConnection()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="resourcer" data-topic="ODBCResourceConnector" data-id="createDBIConnection">}\href{../../resourcer/html/ODBCResourceConnector.html#method-createDBIConnection}{\code{resourcer::ODBCResourceConnector$createDBIConnection()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new DremioResourceConnector instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DremioResourceConnector$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A DremioResourceConnector object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-isFor}{}}}
\subsection{Method \code{isFor()}}{
Check that the provided resource has a URL that locates a Dremio object: the URL scheme must be "odbc+dremio".
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DremioResourceConnector$isFor(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{The resource object to validate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConnectionString"></a>}}
\if{latex}{\out{\hypertarget{method-getConnectionString}{}}}
\subsection{Method \code{getConnectionString()}}{
Get the Dremio ODBC driver connection string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DremioResourceConnector$getConnectionString(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A valid resource object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The Dremio ODBC driver connection string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DremioResourceConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
