% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileResourceClient.R
\docType{class}
\name{FileResourceClient}
\alias{FileResourceClient}
\title{File resource client}
\format{A R6 object of class FileResourceClient}
\usage{
FileResourceClient
}
\description{
Connects to a local file.
}
\keyword{datasets}
