% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileResourceGetter.R
\docType{class}
\name{FileResourceGetter}
\alias{FileResourceGetter}
\title{File resource getter}
\format{A R6 object of class FileResourceGetter}
\usage{
FileResourceGetter
}
\description{
Helper base class for getting the file described by the resource object. ResourceClient class implementations
can use this utility to retrieve files from any locations before processing them according to the declared
data format.
}
\section{Methods}{


\code{$new()} Create new FileResourceGetter instance.
\code{$isFor(resource)} Get a logical that indicates that the file getter is applicable to the provided resource object.
\code{$getFileObject(resource, ...)} Get the file described by the provided resource. Release the connection when done.
}

\keyword{datasets}
