% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileResourceResolver.R
\docType{class}
\name{FileResourceResolver}
\alias{FileResourceResolver}
\title{File Resource resolver}
\format{A R6 object of class FileResourceResolver}
\usage{
FileResourceResolver
}
\description{
The resource is a local file, i.e. which URL scheme is "file" and data format
is one of: csv (comma delimiter), csv2 (semicolon delimiter), tsv (tab delimiter),
spss, sav, por, stata, dta, sas, xpt, excel, xls, xlsx.
}
\section{Methods}{


\code{$new()} Create new FileResourceResolver instance.
\code{$isFor(x)} Get a logical that indicates that the resolver is applicable to the provided resource object.
\code{$newClient()} Make a client for the provided resource.
}

\keyword{datasets}
