% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIResourceConnector.R
\name{findDBIResourceConnector}
\alias{findDBIResourceConnector}
\title{Find a DBI resource connector}
\usage{
findDBIResourceConnector(x)
}
\arguments{
\item{x}{The resource object which corresponding DBI connector is to be found.}
}
\value{
The corresponding DBIResourceConnector object or NULL if none applies.
}
\description{
Find the DBI resource connector that will download the DBI from the provided resource object.
}
