% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileResourceGetter.R
\name{findFileResourceGetter}
\alias{findFileResourceGetter}
\title{Find a file resource getter}
\usage{
findFileResourceGetter(x)
}
\arguments{
\item{x}{The resource object which corresponding file getter is to be found.}
}
\value{
The corresponding FileResourceGetter object or NULL if none applies.
}
\description{
Find the file resource getter that will download the file from the provided resource object.
}
