% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIResourceConnector.R
\name{getDBIResourceConnectors}
\alias{getDBIResourceConnectors}
\title{Get DBI resource connectors registry}
\usage{
getDBIResourceConnectors()
}
\description{
Get the \code{DBIResourceConnector}s registry, create it if it does not exist.
}
