% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileResourceGetter.R
\name{getFileResourceGetters}
\alias{getFileResourceGetters}
\title{Get file resource getters registry}
\usage{
getFileResourceGetters()
}
\description{
Get the \code{FileResourceGetter}s registry, create it if it does not exist.
}
