% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceResolver.R
\name{getResourceResolvers}
\alias{getResourceResolvers}
\title{Get resource resolvers registry}
\usage{
getResourceResolvers()
}
\description{
Get the resource resolvers registry, create it if it does not exist.
}
\examples{
{
resourcer::getResourceResolvers()
}

}
