% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridFsFileResourceGetter.R
\docType{class}
\name{GridFsFileResourceGetter}
\alias{GridFsFileResourceGetter}
\title{GridFS file resource getter}
\format{A R6 object of class GridFsFileResourceGetter}
\usage{
GridFsFileResourceGetter
}
\description{
Access a file that is in the MongoDB file store (GridFS). Credentials may apply.
}
\section{Methods}{


\code{$new(resource)} Create new GridFsFileResourceGetter instance from provided resource object.
\code{$isFor(resource)} Get a logical that indicates that the file getter is applicable to the provided resource object.
\code{$downloadFile(resource)} Get the file described by the provided resource. Release the connection when done.
}

\keyword{datasets}
