% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFileResourceGetter.R
\docType{class}
\name{LocalFileResourceGetter}
\alias{LocalFileResourceGetter}
\title{Local file resource getter}
\format{A R6 object of class LocalFileResourceGetter}
\usage{
LocalFileResourceGetter
}
\description{
Access a file that is stored in the local file system. No credentials apply.
}
\section{Methods}{


\code{$new(resource)} Create new LocalFileResourceGetter instance from provided resource object.
\code{$isFor(resource)} Get a logical that indicates that the file getter is applicable to the provided resource object.
\code{$downloadFile(resource)} Get the file described by the provided resource. Release the connection when done.
}

\keyword{datasets}
