% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{newResource}
\alias{newResource}
\title{Create a Resource}
\usage{
newResource(name = "", url, identity = NULL, secret = NULL,
  clazz = NULL)
}
\arguments{
\item{name}{Otpional human friendly name that identifies the resource.}

\item{url}{URL to access the resource whether it is data or computation capability.}

\item{identity}{User name or account ID (if credentials are applicable).}

\item{secret}{User password or token (if credentials are applicable).}

\item{clazz}{Additional class name, to help resolving data format.}
}
\description{
Creates a new Resource structure.
}
