% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODBCResourceConnector.R
\docType{class}
\name{ODBCResourceConnector}
\alias{ODBCResourceConnector}
\title{ODBC DBI resource connector}
\format{
A R6 object of class ODBCResourceConnector
}
\description{
ODBC DBI resource connector

ODBC DBI resource connector
}
\details{
Makes a ODBC DBI connection from a resource description.
}
\section{Super class}{
\code{\link[resourcer:DBIResourceConnector]{resourcer::DBIResourceConnector}} -> \code{ODBCResourceConnector}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ODBCResourceConnector$new()}}
\item \href{#method-isFor}{\code{ODBCResourceConnector$isFor()}}
\item \href{#method-createDBIConnection}{\code{ODBCResourceConnector$createDBIConnection()}}
\item \href{#method-getConnectionString}{\code{ODBCResourceConnector$getConnectionString()}}
\item \href{#method-clone}{\code{ODBCResourceConnector$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="resourcer" data-topic="DBIResourceConnector" data-id="closeDBIConnection">}\href{../../resourcer/html/DBIResourceConnector.html#method-closeDBIConnection}{\code{resourcer::DBIResourceConnector$closeDBIConnection()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new ODBCResourceConnector instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ODBCResourceConnector$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ODBCResourceConnector object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-isFor}{}}}
\subsection{Method \code{isFor()}}{
Check that the provided resource has a URL that locates a ODBC object: the URL scheme must start with "odbc".
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ODBCResourceConnector$isFor(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{The resource object to validate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createDBIConnection"></a>}}
\if{latex}{\out{\hypertarget{method-createDBIConnection}{}}}
\subsection{Method \code{createDBIConnection()}}{
Creates a DBI connection object from a resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ODBCResourceConnector$createDBIConnection(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A valid resource object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A DBI connection object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConnectionString"></a>}}
\if{latex}{\out{\hypertarget{method-getConnectionString}{}}}
\subsection{Method \code{getConnectionString()}}{
Get the specific ODBC driver connection string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ODBCResourceConnector$getConnectionString(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A valid resource object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The ODBC driver connection string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ODBCResourceConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
