% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalFileResourceGetter.R
\docType{class}
\name{OpalFileResourceGetter}
\alias{OpalFileResourceGetter}
\title{Opal file resource getter}
\format{A R6 object of class OpalFileResourceGetter}
\usage{
OpalFileResourceGetter
}
\description{
Access a file that is stored in a Opal server. Use Basic authentication header if both
resource's identity and secret are defined, token authentication if secret only is defined.
}
\section{Methods}{


\code{$new(resource)} Create new OpalFileResourceGetter instance from provided resource object.
\code{$isFor(resource)} Get a logical that indicates that the file getter is applicable to the provided resource object.
\code{$downloadFile(resource)} Get the file described by the provided resource. Release the connection when done.
}

\keyword{datasets}
