% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoResourceConnector.R
\docType{class}
\name{PrestoResourceConnector}
\alias{PrestoResourceConnector}
\title{Presto DBI resource connector}
\format{A R6 object of class PrestoResourceConnector}
\usage{
PrestoResourceConnector
}
\description{
Makes a Presto DBI connection from a resource description.
}
\section{Methods}{


\code{$new()} Create new PrestoResourceConnector instance.
\code{$isFor(resource)} Get a logical that indicates that the DBI connector is applicable to the provided resource object.
\code{$createDBIConnection(resource, ...)} Get the DBI connection object described by the provided resource.
\code{$closeDBIConnection(conn)} Release the DBI connection when done.
}

\keyword{datasets}
