% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDataFileResourceClient.R
\docType{class}
\name{RDataFileResourceClient}
\alias{RDataFileResourceClient}
\title{R data file resource client}
\format{A R6 object of class TidyFileResourceClient}
\usage{
RDataFileResourceClient
}
\description{
Connects to a R data file and loads it in a controlled environment.
}
\keyword{datasets}
