% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDataFileResourceResolver.R
\docType{class}
\name{RDataFileResourceResolver}
\alias{RDataFileResourceResolver}
\title{R data file Resource resolver}
\format{A R6 object of class RDataFileResourceResolver}
\usage{
RDataFileResourceResolver
}
\description{
The resource is a R data file and data format is the class of the symbol that will be loaded.
}
\section{Methods}{


\code{$new()} Create new TidyFileResourceResolver instance.
\code{$isFor(x)} Get a logical that indicates that the resolver is applicable to the provided resource object.
\code{$newClient()} Make a client for the provided resource.
}

\keyword{datasets}
