% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIResourceConnector.R
\name{registerDBIResourceConnector}
\alias{registerDBIResourceConnector}
\title{Register a DBI resource connector}
\usage{
registerDBIResourceConnector(x)
}
\arguments{
\item{x}{The DBI resource connector object to register.}
}
\description{
Maintain an list of \code{DBIResourceConnector}s that will be called when a new
DBI resource connector is to be found.
}
