% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileResourceGetter.R
\name{registerFileResourceGetter}
\alias{registerFileResourceGetter}
\title{Register a file resource getter}
\usage{
registerFileResourceGetter(x)
}
\arguments{
\item{x}{The file resource getter object to register.}
}
\description{
Maintain an list of \code{FileResourceGetter}s that will be called when a new
file resource getter is to be found.
}
