% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceResolver.R
\name{registerResourceResolver}
\alias{registerResourceResolver}
\title{Register a resource resolver}
\usage{
registerResourceResolver(x)
}
\arguments{
\item{x}{The resource resolver object to register.}
}
\description{
Maintain an list of resource resolvers that will be called when a new
resource is to be resolved.
}
\examples{
\dontrun{
resourcer::registerResourceResolver(MyFileFormatResourceResolver$new())
}

}
