% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resolveResource}
\alias{resolveResource}
\title{Find a resource resolver}
\usage{
resolveResource(x)
}
\arguments{
\item{x}{The resource object which corresponding resolver is to be found.}
}
\description{
Find the resolver that will make a client from the provided resource object.
}
