% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceResolver.R
\docType{class}
\name{ResourceResolver}
\alias{ResourceResolver}
\title{Resource resolver}
\format{A R6 object of class ResourceResolver}
\description{
Helper class for building a Client that implements the access to the data or the
computation unit.
}
\section{Methods}{


\code{$new()} Create new ResourceResolver instance.
\code{$isFor(x)} Get a logical that indicates that the resolver is applicable to the provided resource object.
\code{$newClient()} Make a client for the provided resource.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ResourceResolver$new()}}
\item \href{#method-isFor}{\code{ResourceResolver$isFor()}}
\item \href{#method-newClient}{\code{ResourceResolver$newClient()}}
\item \href{#method-clone}{\code{ResourceResolver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceResolver$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isFor"></a>}}
\subsection{Method \code{isFor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceResolver$isFor(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newClient"></a>}}
\subsection{Method \code{newClient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceResolver$newClient(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
