% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShellResourceClient.R
\docType{class}
\name{ShellResourceClient}
\alias{ShellResourceClient}
\title{Shell resource client}
\format{A R6 object of class ShellResourceClient}
\usage{
ShellResourceClient
}
\description{
Executes local system shell commands.
}
\keyword{datasets}
