% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLResourceClient.R
\docType{class}
\name{SQLResourceClient}
\alias{SQLResourceClient}
\title{SQL database resource client}
\format{A R6 object of class SQLResourceClient}
\usage{
SQLResourceClient
}
\description{
Resource client that connects to a SQL database supported by DBI.
}
\keyword{datasets}
