% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLResourceResolver.R
\docType{class}
\name{SQLResourceResolver}
\alias{SQLResourceResolver}
\title{SQL Database Resource resolver}
\format{A R6 object of class SQLResourceResolver}
\description{
The resource is SQL database.
}
\section{Methods}{


\code{$new()} Create new SQLResourceResolver instance.
\code{$isFor(x)} Get a logical that indicates that the resolver is applicable to the provided resource object.
\code{$newClient()} Make a client for the provided resource.
}

\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{SQLResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-isFor}{\code{SQLResourceResolver$isFor()}}
\item \href{#method-newClient}{\code{SQLResourceResolver$newClient()}}
\item \href{#method-clone}{\code{SQLResourceResolver$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize">}\href{../../resourcer/html/ResourceResolver.html#method-initialize}{\code{resourcer::ResourceResolver$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isFor"></a>}}
\subsection{Method \code{isFor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLResourceResolver$isFor(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newClient"></a>}}
\subsection{Method \code{newClient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLResourceResolver$newClient(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
