% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SshResourceClient.R
\docType{class}
\name{SshResourceClient}
\alias{SshResourceClient}
\title{SSH resource client}
\format{A R6 object of class SshResourceClient}
\usage{
SshResourceClient
}
\description{
Connects to a SSH server.
}
\keyword{datasets}
