% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyFileResourceClient.R
\docType{class}
\name{TidyFileResourceClient}
\alias{TidyFileResourceClient}
\title{Tidy file resource client}
\format{A R6 object of class TidyFileResourceClient}
\usage{
TidyFileResourceClient
}
\description{
Connects to a file and use one of the tidyverse reader.
}
\keyword{datasets}
