% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIResourceConnector.R
\name{unregisterDBIResourceConnector}
\alias{unregisterDBIResourceConnector}
\title{Remove a DBI resource connector from the registry}
\usage{
unregisterDBIResourceConnector(x)
}
\arguments{
\item{x}{The DBI resource connector class name to unregister.}
}
\description{
Remove any DBI resource connectors with the provided class name.
}
