% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileResourceGetter.R
\name{unregisterFileResourceGetter}
\alias{unregisterFileResourceGetter}
\title{Remove a file resource getter from the registry}
\usage{
unregisterFileResourceGetter(x)
}
\arguments{
\item{x}{The file resource getter class name to unregister.}
}
\description{
Remove any file resource getters with the provided class name.
}
