% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceResolver.R
\name{unregisterResourceResolver}
\alias{unregisterResourceResolver}
\title{Remove a resource resolver from the registry}
\usage{
unregisterResourceResolver(x)
}
\arguments{
\item{x}{The resource resolver class name to unregister.}
}
\description{
Remove any resolvers with the provided class name.
}
\examples{
\donttest{
resourcer::unregisterResourceResolver("MyFileFormatResourceResolver")
}

}
