% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HttpFileResourceGetter.R
\docType{class}
\name{HttpFileResourceGetter}
\alias{HttpFileResourceGetter}
\title{HTTP file resource getter}
\format{A R6 object of class HttpFileResourceGetter}
\usage{
HttpFileResourceGetter
}
\description{
Access a file that is stored at a HTTP(S) address. Use Basic authentication header if both resource identity and secret are defined.
}
\section{Methods}{


\code{$new(resource)} Create new HttpFileResourceGetter instance from provided resource object.
\code{$isFor(resource)} Get a logical that indicates that the file getter is applicable to the provided resource object.
\code{$downloadFile(resource)} Get the file described by the provided resource. Release the connection when done.
}

\keyword{datasets}
