% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MariaDBResourceConnector.R
\docType{class}
\name{MariaDBResourceConnector}
\alias{MariaDBResourceConnector}
\title{MariaDB DBI resource connector}
\format{A R6 object of class MariaDBResourceConnector}
\usage{
MariaDBResourceConnector
}
\description{
Makes a MariaDB/MySQL DBI connection from a resource description.
}
\section{Methods}{


\code{$new()} Create new MariaDBResourceConnector instance.
\code{$isFor(resource)} Get a logical that indicates that the DBI connector is applicable to the provided resource object.
\code{$createDBIConnection(resource, ...)} Get the DBI connection object described by the provided resource.
\code{$closeDBIConnection(conn)} Release the DBI connection when done.
}

\keyword{datasets}
