% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{newResource}
\alias{newResource}
\title{Create a Resource}
\usage{
newResource(name = "", url, identity = NULL, secret = NULL, format = NULL)
}
\arguments{
\item{name}{Otpional human friendly name that identifies the resource.}

\item{url}{URL to access the resource whether it is data or computation capability.}

\item{identity}{User name or account ID (if credentials are applicable).}

\item{secret}{User password or token (if credentials are applicable).}

\item{format}{Data format, to help resource resolver identification and coercing to other formats, optional.}
}
\description{
Creates a new Resource structure.
}
\examples{
{
# make a SPSS file resource
res <- resourcer::newResource(
  name = "CNSIM1",
  url = "file:///data/CNSIM1.sav",
  format = "spss"
)
}

}
