% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoSQLResourceClient.R
\docType{class}
\name{NoSQLResourceClient}
\alias{NoSQLResourceClient}
\title{NoSQL database resource client}
\format{A R6 object of class NoSQLResourceClient}
\usage{
NoSQLResourceClient
}
\description{
Resource client that connects to a NoSQL database supported by nodbi.
}
\keyword{datasets}
