% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoSQLResourceResolver.R
\docType{class}
\name{NoSQLResourceResolver}
\alias{NoSQLResourceResolver}
\title{NoSQL Database Resource resolver}
\format{
A R6 object of class NoSQLResourceResolver
}
\description{
NoSQL Database Resource resolver

NoSQL Database Resource resolver
}
\details{
The resource is NoSQL database such as mongodb (elasticsearch, redis, couchdb, sqlite are not supported yet).
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{NoSQLResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-isFor}{\code{NoSQLResourceResolver$isFor()}}
\item \href{#method-newClient}{\code{NoSQLResourceResolver$newClient()}}
\item \href{#method-clone}{\code{NoSQLResourceResolver$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize">}\href{../../resourcer/html/ResourceResolver.html#method-initialize}{\code{resourcer::ResourceResolver$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-isFor}{}}}
\subsection{Method \code{isFor()}}{
Check that the provided resource has a URL that locates a nodbi object: the URL scheme must be one of "mongodb", "mongodb+srv". Other NoSQL databases "elasticsearch", "redis", "couchdb", "sqlite" are not supported yet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoSQLResourceResolver$isFor(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The resource object to validate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-newClient}{}}}
\subsection{Method \code{newClient()}}{
Creates a NoSQLResourceClient instance from provided resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoSQLResourceResolver$newClient(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A valid resource object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A NoSQLResourceClient object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoSQLResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
