% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostgresResourceConnector.R
\docType{class}
\name{PostgresResourceConnector}
\alias{PostgresResourceConnector}
\title{Postgres DBI resource connector}
\format{A R6 object of class PostgresResourceConnector}
\usage{
PostgresResourceConnector
}
\description{
Makes a Postgres DBI connection from a resource description.
}
\section{Methods}{


\code{$new()} Create new PostgresResourceConnector instance.
\code{$isFor(resource)} Get a logical that indicates that the DBI connector is applicable to the provided resource object.
\code{$createDBIConnection(resource, ...)} Get the DBI connection object described by the provided resource.
\code{$closeDBIConnection(conn)} Release the DBI connection when done.
}

\keyword{datasets}
