% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceClient.R
\docType{class}
\name{ResourceClient}
\alias{ResourceClient}
\title{Resource client}
\format{A R6 object of class ResourceClient}
\usage{
ResourceClient
}
\description{
Helper class for connecting to a resource data store or a computation unit.
}
\section{Methods}{


\code{$new(resource)} Create new ResourceClient instance from provided resource object.
\code{$getResource()} Get the resource reference object.
\code{$getConnection()} Get the raw connection object to the resource.
\code{$downloadFile(fileext)} Coerce the resource to a file with provided file extension.
\code{$asDataFrame()} Coerce the resource to a data.frame.
\code{$close()} Close the connection with the resource.
}

\keyword{datasets}
