% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceResolver.R
\docType{class}
\name{ResourceResolver}
\alias{ResourceResolver}
\title{Resource resolver}
\format{A R6 object of class ResourceResolver}
\usage{
ResourceResolver
}
\description{
Helper class for building a Client that implements the access to the data or the
computation unit.
}
\section{Methods}{


\code{$new()} Create new ResourceResolver instance.
\code{$isFor(x)} Get a logical that indicates that the resolver is applicable to the provided resource object.
\code{$newClient()} Make a client for the provided resource.
}

\keyword{datasets}
