% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3FileResourceGetter.R
\docType{class}
\name{S3FileResourceGetter}
\alias{S3FileResourceGetter}
\title{AWS S3 file resource getter}
\format{A R6 object of class S3FileResourceGetter}
\usage{
S3FileResourceGetter
}
\description{
Access a file that is in the Amazon Web Services S3 file store. The host name is the bucket name. Credentials may apply.
}
\section{Methods}{


\code{$new(resource)} Create new S3FileResourceGetter instance from provided resource object.
\code{$isFor(resource)} Get a logical that indicates that the file getter is applicable to the provided resource object.
\code{$downloadFile(resource)} Get the file described by the provided resource. Release the connection when done.
}

\keyword{datasets}
