% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScpFileResourceGetter.R
\docType{class}
\name{ScpFileResourceGetter}
\alias{ScpFileResourceGetter}
\title{SCP file resource getter}
\format{A R6 object of class ScpFileResourceGetter}
\usage{
ScpFileResourceGetter
}
\description{
Access a file that is stored on a server accessible through SSH. Credentials apply.
}
\section{Methods}{


\code{$new(resource)} Create new ScpFileResourceGetter instance from provided resource object.
\code{$isFor(resource)} Get a logical that indicates that the file getter is applicable to the provided resource object.
\code{$downloadFile(resource)} Get the file described by the provided resource. Release the connection when done.
}

\keyword{datasets}
