% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShellResourceResolver.R
\docType{class}
\name{ShellResourceResolver}
\alias{ShellResourceResolver}
\title{Shell Resource resolver}
\format{A R6 object of class ShellResourceResolver}
\usage{
ShellResourceResolver
}
\description{
The resource is a computation unit, accessible by issuing local system commands, i.e. which URL scheme is "sh".
}
\section{Methods}{


\code{$new()} Create new ShellResourceResolver instance.
\code{$isFor(x)} Get a logical that indicates that the resolver is applicable to the provided resource object.
\code{$newClient()} Make a client for the provided resource.
}

\keyword{datasets}
