% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparkResourceConnector.R
\docType{class}
\name{SparkResourceConnector}
\alias{SparkResourceConnector}
\title{Spark DBI resource connector}
\format{A R6 object of class SparkResourceConnector}
\usage{
SparkResourceConnector
}
\description{
Makes a Spark connection object, that is also a DBI connection object, from a resource description.
}
\section{Methods}{


\code{$new()} Create new SparkResourceConnector instance.
\code{$isFor(resource)} Get a logical that indicates that the DBI connector is applicable to the provided resource object.
\code{$createDBIConnection(resource, ...)} Get the DBI connection object described by the provided resource.
\code{$closeDBIConnection(conn)} Release the DBI connection when done.
}

\keyword{datasets}
