% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLResourceResolver.R
\docType{class}
\name{SQLResourceResolver}
\alias{SQLResourceResolver}
\title{SQL Database Resource resolver}
\format{A R6 object of class SQLResourceResolver}
\usage{
SQLResourceResolver
}
\description{
The resource is SQL database.
}
\section{Methods}{


\code{$new()} Create new SQLResourceResolver instance.
\code{$isFor(x)} Get a logical that indicates that the resolver is applicable to the provided resource object.
\code{$newClient()} Make a client for the provided resource.
}

\keyword{datasets}
