% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLResourceResolver.R
\docType{class}
\name{SQLResourceResolver}
\alias{SQLResourceResolver}
\title{SQL Database Resource resolver}
\format{
A R6 object of class SQLResourceResolver
}
\description{
SQL Database Resource resolver

SQL Database Resource resolver
}
\details{
The resource is SQL database.
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{SQLResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SQLResourceResolver-isFor}{\code{SQLResourceResolver$isFor()}}
\item \href{#method-SQLResourceResolver-newClient}{\code{SQLResourceResolver$newClient()}}
\item \href{#method-SQLResourceResolver-clone}{\code{SQLResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-SQLResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
Check that the provided resource has a registered DBIResourceConnector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLResourceResolver$isFor(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The resource object to evaluate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-SQLResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
Creates a SQLResourceClient instance from provided resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLResourceResolver$newClient(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A valid resource object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A SQLResourceClient object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SQLResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
