% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SshResourceResolver.R
\docType{class}
\name{SshResourceResolver}
\alias{SshResourceResolver}
\title{SSH Resource resolver}
\format{A R6 object of class SshResourceResolver}
\usage{
SshResourceResolver
}
\description{
The resource is a computation unit, accessible through SSH, i.e. which URL scheme is "ssh".
}
\section{Methods}{


\code{$new()} Create new SshResourceResolver instance.
\code{$isFor(x)} Get a logical that indicates that the resolver is applicable to the provided resource object.
\code{$newClient()} Make a client for the provided resource.
}

\keyword{datasets}
